+++ lib/membrane/core/element/manual_flow_controller.ex
@@ -200,8 +200,19 @@ defmodule Membrane.Core.Element.ManualFlowController do
          {:buffers, buffers, _inbound_metric_buf_size, outbound_metric_buf_size},
          state
        ) do
+    demand_unit = PadModel.get_data!(state, pad_ref, :demand_unit)
+
     state =
-      PadModel.update_data!(state, pad_ref, :manual_demand_size, &(&1 - outbound_metric_buf_size))
+      PadModel.update_data!(
+        state,
+        pad_ref,
+        :manual_demand_size,
+        &Membrane.Core.Element.ManualFlowController.BufferMetric.reduce_demand(
+          demand_unit,
+          &1,
+          outbound_metric_buf_size
+        )
+      )
 
     BufferController.exec_buffer_callback(pad_ref, buffers, state)
   end