+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * Update required Elixir version to `~> 1.17`
 * Add a Mix task for generating demos [#1067](https://github.com/membraneframework/membrane_core/pull/1067)
 * Add Mix tasks for easier initialization of components [#1059](https://github.com/membraneframework/membrane_core/pull/1059)
+* Support input demand unit expressed in timestamps [#1079](https://github.com/membraneframework/membrane_core/pull/1079)
 
 ## 1.2.6
  * Add tutorials [#1007](https://github.com/membraneframework/membrane_core/pull/1007), plugins [#1012](https://github.com/membraneframework/membrane_core/pull/1012) and demos [#1013](https://github.com/membraneframework/membrane_core/pull/1013) to the docs.